refactor(HELP): Removed code for old help system and tests that helped with the transitions
