+++ src/args/arg_builder/flag.rs
@@ -1,14 +1,13 @@
 // use std::collections::HashSet;
 use std::fmt::{Display, Formatter, Result};
 use std::convert::From;
-use std::io;
 use std::rc::Rc;
 use std::result::Result as StdResult;
 
 use vec_map::VecMap;
 
 use Arg;
-use args::{AnyArg, HelpWriter, DispOrder};
+use args::{AnyArg, DispOrder};
 use args::settings::{ArgFlags, ArgSettings};
 
 #[derive(Debug)]
@@ -47,11 +46,6 @@ impl<'n, 'e> FlagBuilder<'n, 'e> {
             name: name,
             ..Default::default()
         }
-    }
-
-    pub fn write_help<W: io::Write>(&self, w: &mut W, longest: usize, nlh: bool) -> io::Result<()> {
-        let hw = HelpWriter::new(self, longest, nlh);
-        hw.write_to(w)
     }
 }
