+++ src/args/arg_builder/positional.rs
@@ -1,12 +1,11 @@
 use std::fmt::{Display, Formatter, Result};
 use std::result::Result as StdResult;
 use std::rc::Rc;
-use std::io;
 
 use vec_map::VecMap;
 
 use Arg;
-use args::{AnyArg, HelpWriter, DispOrder};
+use args::{AnyArg, DispOrder};
 use args::settings::{ArgFlags, ArgSettings};
 
 #[allow(missing_debug_implementations)]
@@ -105,11 +104,6 @@ impl<'n, 'e> PosBuilder<'n, 'e> {
         pb
     }
 
-    pub fn write_help<W: io::Write>(&self, w: &mut W, longest: usize, skip_pv: bool, nlh: bool) -> io::Result<()> {
-        let mut hw = HelpWriter::new(self, longest, nlh);
-        hw.skip_pv = skip_pv;
-        hw.write_to(w)
-    }
 }
 
 impl<'n, 'e> Display for PosBuilder<'n, 'e> {
